requirements.txt
Mastodon.py==2.0.1
pandas
